+++ desktop/src/preload/api/transfer/parseXlsxFileList.ts
@@ -3,73 +3,61 @@ import { isAccessionValid } from "./isAccessionValid";
 import { isApplicationValid } from "./isApplicationValid";
 
 export const parseXlsxFileList = (
-  fileList: File | null | undefined
+  file: File | null | undefined
 ): Promise<{
   accession: string;
   application: string;
   folders: string[];
 } | null> => {
   return new Promise((resolve, reject) => {
-    if (fileList) {
-      const reader = new FileReader();
+    if (!file) return reject("Missing filelist.");
 
-      reader.onload = (event) => {
-        try {
-          if (event.target?.result) {
-            const data = new Uint8Array(event.target.result as ArrayBuffer);
-            const workbook = XLSX.read(data, { type: "array" });
-            const coverPage = workbook.Sheets["COVER PAGE"];
-            const fileList = workbook.Sheets["FILE LIST"];
+    const reader = new FileReader();
 
-            if (coverPage && fileList) {
-              // Access specific cells
-              const accessionCell = coverPage.B3?.v; // B3 cell value
-              const applicationCell = coverPage.B4?.v; // B4 cell value
+    reader.onload = (event) => {
+      try {
+        if (event.target?.result) {
+          const data = new Uint8Array(event.target.result as ArrayBuffer);
+          const workbook = XLSX.read(data, { type: "array" });
+          const coverPage = workbook.Sheets["COVER PAGE"];
+          const fileList = workbook.Sheets["FILE LIST"];
 
-              // Extract folder names from column A starting at A2
-              const folders: string[] = [];
-              let rowIndex = 2; // Start at row 2 (A2)
-              while (true) {
-                const cellAddress = `A${rowIndex}`;
-                const cell = fileList[cellAddress];
-                if (cell?.v) {
-                  folders.push(cell.v.toString());
-                  rowIndex++;
-                } else {
-                  break; // Stop when an empty cell is encountered
-                }
-              }
+          if (!coverPage || !fileList)
+            reject('Missing on or more of ["COVER PAGE", "FILE LIST"].');
 
-              if (
-                accessionCell &&
-                applicationCell &&
-                isAccessionValid(accessionCell.toString()) &&
-                isApplicationValid(applicationCell.toString())
-              ) {
-                resolve({
-                  accession: accessionCell.toString(),
-                  application: applicationCell.toString(),
-                  folders,
-                });
-              } else {
-                resolve(null); // One or both cells are empty
-              }
+          // Access cells
+          const accession = coverPage.B3?.v.toString();
+          const application = coverPage.B4?.v.toString();
+
+          // Extract folder names from column A starting at A2
+          const folders: string[] = [];
+          let rowIndex = 2; // Start at row 2 (A2)
+          while (true) {
+            const cellAddress = `A${rowIndex}`;
+            const cell = fileList[cellAddress];
+            if (cell?.v) {
+              folders.push(cell.v.toString());
+              rowIndex++;
             } else {
-              reject(new Error("Sheet not found in the Excel file."));
+              break; // Stop when an empty cell is encountered
             }
           }
-        } catch (error) {
-          reject(error);
+
+          if (!(isAccessionValid(accession) && isApplicationValid(application)))
+            reject("Invalid accession and/or application.");
+
+          resolve({
+            accession,
+            application,
+            folders,
+          });
         }
-      };
+      } catch (error) {
+        reject(`${error}`);
+      }
+    };
 
-      reader.onerror = () => {
-        reject(new Error("Failed to read the file."));
-      };
-
-      reader.readAsArrayBuffer(fileList);
-    } else {
-      resolve(null);
-    }
+    reader.onerror = () => reject("Failed to read the file.");
+    reader.readAsArrayBuffer(file);
   });
 };
