+++ desktop/src/renderer/src/components/transfer/UploadGrid/index.ts
@@ -1,0 +1,1 @@
+export * from "./FolderUploadGrid";
